refactor(model): split models into domain package
- monolithic models.py mixed every domain and grew on each tool addition
- group into common/projects/documents/work_items/links/comments, re-export from package init
